Integration fixes. NonPersistent Neo4 store.

+++ ab-engine/src/main/java/com/auditbucket/engine/repo/neo4j/AuditDaoNeo.java
@@ -70,22 +70,20 @@ public class AuditDaoNeo implements AuditDao {
 
     private Logger logger = LoggerFactory.getLogger(AuditDaoNeo.class);
 
-    @Override
-    @Caching(evict = {@CacheEvict(value = "auditHeaderId", key = "#p0.id"),
-            @CacheEvict(value = "auditKey", key = "#p0.auditKey")})
     public AuditHeader save(AuditHeader auditHeader) {
         return save(auditHeader, null);
     }
+
     @Override
     @Caching(evict = {@CacheEvict(value = "auditHeaderId", key = "#p0.id"),
-                     @CacheEvict(value = "auditKey", key = "#p0.auditKey")}
+            @CacheEvict(value = "auditKey", key = "#p0.auditKey")}
     )
     public AuditHeader save(AuditHeader auditHeader, DocumentType documentType) {
         auditHeader.bumpUpdate();
         AuditHeader header = auditRepo.save((AuditHeaderNode) auditHeader);
-        if ( header!=null && documentType !=null ) {
+        if (header != null && documentType != null) {
             Node node = template.getPersistentState(header);
-            node.addLabel( DynamicLabel.label(documentType.getName()));
+            node.addLabel(DynamicLabel.label(documentType.getName()));
         }
 
         return header;
@@ -144,8 +142,8 @@ public class AuditDaoNeo implements AuditDao {
     @Override
     public Collection<AuditHeader> findHeaders(Long fortressId, String label, Long skipTo) {
         //ToDo: Should this pass in timestamp it got to??
-        String cypher = "match (f:Fortress)-[:TRACKS]->(audit:`"+label+"`) where id(f)={fortress} return audit ORDER BY audit.dateCreated ASC skip {skip} limit 100 ";
-        Map<String,Object>args = new HashMap<>();
+        String cypher = "match (f:Fortress)-[:TRACKS]->(audit:`" + label + "`) where id(f)={fortress} return audit ORDER BY audit.dateCreated ASC skip {skip} limit 100 ";
+        Map<String, Object> args = new HashMap<>();
         args.put("fortress", fortressId);
         args.put("skip", skipTo);
         Result<Map<String, Object>> result = template.query(cypher, args);
@@ -171,20 +169,20 @@ public class AuditDaoNeo implements AuditDao {
     }
 
     @Override
-    public TxRef findTxTag(@NotEmpty String userTag, @NotNull Company company, boolean fetchHeaders) {
-        return auditRepo.findTxTag(userTag, company.getId());
+    public TxRef findTxTag(@NotEmpty String txTag, @NotNull Company company, boolean fetchHeaders) {
+        return auditRepo.findTxTag(txTag, company.getId());
     }
 
 
     @Override
     public TxRef beginTransaction(String id, Company company) {
 
-        TxRef tag = findTxTag(id, company, false);
-        if (tag == null) {
-            tag = new TxRefNode(id, company);
-            template.save(tag);
+        TxRef txTag = findTxTag(id, company, false);
+        if (txTag == null) {
+            txTag = new TxRefNode(id, company);
+            template.save(txTag);
         }
-        return tag;
+        return txTag;
     }
 
     @Override
@@ -235,13 +233,8 @@ public class AuditDaoNeo implements AuditDao {
         return result;
     }
 
-    @Override
-    public AuditLog addLog(AuditHeader header, AuditChange al, DateTime fortressWhen) {
-        return template.save(new AuditLogRelationship(header, al, fortressWhen));
-
-    }
-
     public AuditLog save(AuditLog log) {
+        logger.debug("Saving audit log [{}] - AuditChange ID [{}]", log, log.getAuditChange().getId());
         return template.save((AuditLogRelationship) log);
     }
 
@@ -252,7 +245,7 @@ public class AuditDaoNeo implements AuditDao {
 //        Map<String, Object> ab = new HashMap<>();
 //        ab.put("name", "AuditBucket");
 //        ArrayList labels = new ArrayList();
-//        labels.add("ZZTest");
+//        labels.add("ABPing");
 //        Node abNode = template.getGraphDatabase().getOrCreateNode("system", "name", "ab", "AuditBucket", ab, labels);
 //        if (abNode == null) {
 //            return "Neo4J has problems";
@@ -285,7 +278,7 @@ public class AuditDaoNeo implements AuditDao {
     }
 
     @Override
-    @Cacheable(value = "auditLog")
+    @Cacheable(value = "auditLog", unless = "#result==null")
     public AuditLog getLog(Long logId) {
         Relationship change = template.getRelationship(logId);
         if (change != null)
@@ -308,43 +301,82 @@ public class AuditDaoNeo implements AuditDao {
         return template.fetch(lastChange);
     }
 
-    @Override
-    public AuditLog getLastLog(Long headerId) {
-        return auditLogRepo.getLastAuditLog(headerId);
-    }
-
     enum LastChange implements RelationshipType {
         LAST_CHANGE
     }
+
     private static final String LAST_CHANGE = "LAST_CHANGE";
 
-    public AuditLog getLastAuditLog(Long auditHeaderID) {
-
-        AuditLog when = auditLogRepo.getLastAuditLog(auditHeaderID);
-        if (when != null) {
-            logger.trace("Last Change {}", when);
+    @Override
+    public AuditLog addLog(AuditHeader auditHeader, AuditChange newChange, DateTime fortressWhen, AuditLog existingLog) {
+        AuditLog newLog = template.save(new AuditLogRelationship(auditHeader, newChange, fortressWhen));
+        boolean moreRecent = (existingLog == null || existingLog.getFortressWhen() <= newLog.getFortressWhen());
+        if (moreRecent) {
+            removeLastChange(auditHeader, existingLog);
+            makeLastChange(auditHeader, newChange);
         }
 
-        return when;
+        logger.debug("Added Log - AuditHeader [{}], Log [{}], Change [{}]", auditHeader.getId(), newLog, newChange.getId());
+        return newLog;
+
+    }
+
+    private void removeLastChange(AuditHeader auditHeader, AuditLog existingLog) {
+        if (existingLog != null) {
+            Node auditNode = template.getPersistentState(auditHeader);
+            if (existingLog.getAuditChange() != null) {
+                Node changeNode = template.getPersistentState(existingLog.getAuditChange());
+                Relationship r = template.getRelationshipBetween(auditNode, changeNode, LAST_CHANGE);
+                if (r != null) {
+                    logger.debug("removeLastChange AuditHeader[{}], [{}]", auditHeader.getId(), r);
+                    r.delete();
+                    r = auditNode.getSingleRelationship(LastChange.LAST_CHANGE, Direction.OUTGOING);
+                    if (r != null) r.delete();
+                } else {
+                    logger.debug("No change to remove for AuditHeader [{}]", auditHeader.getId());
+                }
+            }
+        }
     }
 
     @Override
-    public void setLastChange(AuditHeader auditHeader, AuditChange toAdd) {
+    public void makeLastChange(AuditHeader auditHeader, AuditChange lastChange) {
+        Node auditNode = template.getPersistentState(auditHeader);
+        Node changeNode = template.getPersistentState(lastChange);
+        Relationship r = template.createRelationshipBetween(auditNode, changeNode, LAST_CHANGE, null);
+        logger.debug("makeLastChange - AuditHeader [{}], LAST_CHANGE [{}], auditChange [{}]", auditHeader.getId(), r.getId(), changeNode.getId());
+    }
 
-        Node header = template.getPersistentState(auditHeader);
-        Relationship r = header.getSingleRelationship(LastChange.LAST_CHANGE, Direction.OUTGOING);
+
+    public AuditLog getLastAuditLog(Long auditHeaderID) {
+        AuditLogRelationship log = null;
+
+        Relationship r = template.getNode(auditHeaderID).getSingleRelationship(LastChange.LAST_CHANGE, Direction.BOTH);
         if (r != null) {
-            logger.info("Deleting {} for {}",r, auditHeader.getCallerRef());
-            r.delete();
-        }
+            auditLogRepo.getLastAuditLog(auditHeaderID);
 
-        Node auditChange = template.getPersistentState(toAdd);
-        template.createRelationshipBetween(header, auditChange, LAST_CHANGE, null);
-        logger.info("Created {}, for {}" ,header.getSingleRelationship(LastChange.LAST_CHANGE, Direction.OUTGOING), auditHeader.getCallerRef());
+
+            log = auditLogRepo.getLastAuditLog(r.getEndNode().getId());
+//            int count = 0;
+//
+//            for (AuditLogRelationship when : logs) {
+//                count++;
+//                log = when;
+//                logger.debug("getLastLog for AuditHeader [{}], found RLX [{}], changeId [{}]", auditHeaderID, when, when.getAuditChange().getId());
+//            }
+//
+//            if (log != null) {
+//                if (count > 1)
+//                    logger.debug("*** AuditHeader [{}] - Found more than [{}] LAST_CHANGE Log, should only be 1", auditHeaderID, count);
+//
+//            }
+        }
+        return log;
     }
 
     @Override
     public AuditChange save(AuditChange change, Boolean compressed) {
+        logger.debug("Saving Audit Change [{}]", change);
         return template.save(change);
     }
